Calculate the sub-totals of the purchase-/sales order in the default currency
Instead of summing all prices in the table which could be in different currencies,
load them through a callback and the API.

TODO:
The onPostFooter callback is called twice per tabel on load, no idea why. Therefore
4 calls to the API are being done at page load.
Still looking for a better way to reduce the amount of API calls so this is still a
work in progres.