Reverted previous attempt on OrderSerializer and created a new version
The previous attemt did various API calls and JS callback to achieve the desired result.
This attempt works more along the lines of the original code. The various serializers for
the PO/SO item and extra line are adding pre-converted currencies.
Therefore the original JS footer code can do it's job but with converted values.
Also the update of the Total price field wasn't correct preventing the update of the total price.